# src/swagger.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Swagger Contacts API",
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "http://localhost:4000"
		}
	],
...
	"paths": {
...
		"/recipes/{id}": {
			"get": {
				"tags": ["Recipes"],
				"description": "Get Recipe by id",
				"summary": "Return recipe with given id",
				"operationId": "getRecipeById",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "pass recipe id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {
											"type": "object",
											"properties": {
												"status": {
													"type": "integer",
													"default": 200
												},
												"message": {
													"type": "string",
													"default": "Success"
												},
												"recipe": {
													"$ref": "#/components/schemas/Recipe"
												}
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/BadRequest"
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"404": {
						"$ref": "#/components/responses/RecipeNotFoundError"
					}
				},
				"security": [
					{
						"jwtAuth": []
					}
				]
			}
		},
		"/recipes/category/{category}": {
			"get": {
				"tags": ["Recipes"],
				"summary": "Get recipes by category",
				"description": "Return 8 recipes with given category for each page",
				"operationId": "getRecipesByCategory",
				"parameters": [
					{
						"name": "category",
						"in": "path",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Category"
						}
					},
					{
						"name": "page",
						"in": "query",
						"description": "pass the page number",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "A list of recipes in category",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecipeSmalResult"
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/BadRequest"
					},
					"401": {
						"$ref": "#/components/responses/UnauthorizedError"
					},
					"404": {
						"description": "Category not found!",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"status": {
											"type": "integer",
											"example": 404
										},
										"message": {
											"type": "string",
											"example": "Category not found!"
										}
									}
								}
							}
						}
					}
				},
				"security": [
					{
						"jwtAuth": []
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"RecipeSmalResult": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"title": {
						"type": "string",
						"example": "Macaroon"
					},
					"instructions": {
						"type": "string",
						"example": "Put 2 tablespoons of the oil in a large pot over medium heat. When it’s hot, add onion, carrot, celery and garlic; sprinkle with salt and pepper and cook, stirring occasionally, until vegetables are soft, 5 to 10 minutes...."
					},
					"preview": {
						"type": "string"
					}
				}
			},
...
			"Recipe": {
				"type": "object",
				"required": ["recipeTitle", "description", "category", "cookingTime", "ingredients", "preparation", "image"],
				"properties": {
					"_id": {
						"type": "string"
					},
					"title": {
						"type": "string",
						"example": "Macaroon"
					},
					"description": {
						"type": "string",
						"example": "it's an awesome cookies"
					},
					"category": {
						"type": "string",
						"example": "dessert"
					},
					"cookingTime": {
						"type": "string",
						"default": "2 hours"
					},
					"favorites": {
						"type": "array",
						"items": {
							"type": "string",
							"default": "user id's"
						}
					},
					"popularity": {
						"description": "Depends on how many users added it to favorite",
						"type": "integer",
						"default": 0
					},
					"ingredients": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/RecipeIngredients"
						}
					},
					"preparation": {
						"type": "string",
						"example": "beat eggs until smooth"
					},
					"image": {
						"type": "string",
						"format": "binary"
					},
					"owner": {
						"type": "string",
						"example": "Owner's id"
					}
				}
			},
...
			"Category": {
				"type": "string",
				"enum": [
					"Beef",
					"Breakfast",
					"Dessert",
					"Goat",
					"Lamb",
					"Miscellaneous",
					"Vegan",
					"Chicken",
					"Pasta",
					"Pork",
					"SeaFood",
					"Side"
				]
			},
			"RecipeIngredient": {
				"type": "object",
				"required": ["name", "measure"],
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"example": "eggs"
					},
					"measure": {
						"type": "string",
						"example": "2 tbs"
					},
					"image": {
						"type": "string",
						"example": "https://res.cloudinary.com/ddbvbv5sp/image/upload/v1678564856/ppyhvetrcdcexkwcgaj3.png"
					},
					"family": {
						"type": "string"
					}
				}
			},
			"RecipeIngredients": {
				"type": "array",
				"items": {
					"$ref": "#/components/schemas/RecipeIngredient"
				}
			},
...
		},
		"responses": {
...
			"RecipeNotFoundError": {
				"description": "Recipe not found",
				"content": {
					"application/json": {
						"schema": {
							"type": "object",
							"properties": {
								"status": {
									"type": "integer",
									"example": 404
								},
								"message": {
									"type": "string",
									"example": "Recipe not found!"
								}
							}
						}
					}
				}
			},
...
			"UnauthorizedError": {
				"description": "Access token is missing or invalid",
				"content": {
					"application/json": {
						"schema": {
							"type": "object",
							"properties": {
								"status": {
									"type": "integer",
									"example": 401
								},
								"message": {
									"type": "string",
									"example": "Access token is missing or invalid"
								}
							}
						}
					}
				}
			},
			"BadRequest": {
				"description": "Bad request!",
				"content": {
					"application/json": {
						"schema": {
							"type": "object",
							"properties": {
								"status": {
									"type": "integer",
									"example": 400
								},
								"message": {
									"type": "string",
									"example": "Bad request!"
								}
							}
						}
					}
				}
			},
...
		},
		"securitySchemes": {
			"jwtAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"security": [
		{
			"jwtAuth": []
		}
	]
}